requirements.txt
streamlit
pandas
pdfplumber
pytesseract
pymupdf
Pillow
reportlab